fix tab_control and tab_page

+++ src/xtd_forms/src/xtd/forms/tab_page.cpp
@@ -31,12 +31,7 @@ control& tab_page::text(const ustring& text) {
   return control::text(text);
 }
 
-void tab_page::create_handle() {
-  panel::create_handle();
-  native::tab_control::add_item(this->parent().value().get().handle(), this->handle_, this->text_);
-}
-
 void tab_page::destroy_handle() {
-  native::tab_control::delete_item(this->parent().value().get().handle(), this->handle_);
+  if (this->parent().has_value()) native::tab_control::delete_item(this->parent().value().get().handle(), this->handle_);
   panel::destroy_handle();
 }
